ci(dco): added dco file
By adding the Developer Certificate of Origin to our repository, we can add the DCO enforcer application from GitHub. This is a standard practice in Open Source and ensures that contributors have to sign off that they understand that they are contributing to open source, and do not own any of the code that they supply.

Signed-off-by: Roberth Strand <rob@azure.tf>